+++ src/Linear/Logic.hs
@@ -40,7 +40,7 @@
 -- To use the alias for @('*')@, make sure to enable @{-# LANGUAGE NoStarIsType #-}@
 
 module Linear.Logic
-( Prep, Prop'(Not,(!=)), Prop
+( Prep, Prop(Not,(!=),(=!))
 -- additive conjunction, with
 , type (&)(..), Top(..), type With, with, withL', withR', withL, withR
 -- additive disjunction, oplus