+++ src/Linear/Logic/Day.hs
@@ -1,4 +1,3 @@
-
 {-# language LinearTypes #-}
 {-# language RankNTypes #-}
 {-# language LambdaCase #-}
@@ -28,18 +27,18 @@ import Prelude.Linear ((&))
 
 -- | Day convolution of logical functors
 data Day f g a where
-  Day :: (Prop' b, Prop' c) => ((b,c) ⊸ a) %1 -> f b %1 -> g c %1 -> Day f g a
+  Day :: (Prop b, Prop c) => ((b,c) ⊸ a) %1 -> f b %1 -> g c %1 -> Day f g a
 
 -- | refuted Day convolution of logical functors
 newtype Night f g a = Night 
-  ( forall b c. (Prop' b, Prop' c) =>
+  ( forall b c. (Prop b, Prop c) =>
     (a <#- (b,c)) ⅋ Not (f b) ⅋ Not (g c)
   )
 
-instance (Functor f, Functor g, Prop' a) => Prop' (Day f g a) where
+instance (Functor f, Functor g, Prop a) => Prop (Day f g a) where
   type Not (Day f g a) = Night f g a 
   Day bca (fb :: f b) gc != Night no = (bca,(fb,gc)) != no
 
-instance (Functor f, Functor g, Prop' a) => Prop' (Night f g a) where
+instance (Functor f, Functor g, Prop a) => Prop (Night f g a) where
   type Not (Night f g a) = Day f g a
   Night no != Day bca (fb :: f b) gc = (bca,(fb,gc)) != no