+++ src/Linear/Logic/Yoneda.hs
@@ -12,7 +12,7 @@
 {-# language NoImplicitPrelude #-}
 {-# language TypeFamilies #-}
 {-# language TypeApplications #-}
-{-# options_ghc -Wno-unused-imports #-}
+{-# options_ghc -Wno-unused-imports -fplugin Linear.Logic.Plugin #-}
 
 -- | Yoneda and Coyoneda constructions for linear-logic functors.
 module Linear.Logic.Yoneda
@@ -36,29 +36,29 @@ import Linear.Logic.Functor
 import Linear.Logic.Y
 
 -- | Yoneda encoding for a linear-logic 'Functor'.
-newtype Yoneda f a = Yoneda (forall r. Prop' r => Ur (a ⊸ r) ⊸ f r)
+newtype Yoneda f a = Yoneda (forall r. Prop r => Ur (a ⊸ r) ⊸ f r)
 
 -- | Refutations of 'Yoneda'.
 data Noneda f a where
-  Noneda :: Prop' r => {-# unpack #-} !(f r <#- Ur (a ⊸ r)) %1 -> Noneda f a
+  Noneda :: Prop r => {-# unpack #-} !(f r <#- Ur (a ⊸ r)) %1 -> Noneda f a
 
-instance (Functor f, Prop' a) => Prop' (Yoneda f a) where
+instance (Functor f, Prop a) => Prop (Yoneda f a) where
   type Not (Yoneda f a) = Noneda f a
   Yoneda y != Noneda n = y != n
 
-instance (Functor f, Prop' a) => Prop' (Noneda f a) where
+instance (Functor f, Prop a) => Prop (Noneda f a) where
   type Not (Noneda f a) = Yoneda f a
   Noneda y != Yoneda n = y != n
 
-runYoneda :: Prop' r => Yoneda f a %1 -> (a ⊸ r) -> f r
+runYoneda :: Prop r => Yoneda f a %1 -> (a ⊸ r) -> f r
 runYoneda (Yoneda f) a2r = fun' f (Ur a2r)
 
-lowerYoneda' :: (Prop' a, Lol l) => l (Yoneda f a) (f a)
+lowerYoneda' :: (Prop a, Lol l) => l (Yoneda f a) (f a)
 lowerYoneda' = lol \case
   L -> \nfa -> Noneda (Ur id :-#> nfa)
   R -> \f -> runYoneda f id
 
-liftYoneda :: forall f a i. (Functor f, Prop' a, Iso i) => i (f a) (Yoneda f a)
+liftYoneda :: forall f a i. (Functor f, Prop a, Iso i) => i (f a) (Yoneda f a)
 liftYoneda = iso \case
   L -> lowerYoneda'
   R -> lol \case
@@ -68,28 +68,28 @@ liftYoneda = iso \case
         R -> \f -> fmap' f fa
         L -> \nfr -> whyNot \a2r -> fmap a2r fa != nfr
 
-lowerYoneda :: forall f a i. (Functor f, Prop' a, Iso i) => i (Yoneda f a) (f a)
+lowerYoneda :: forall f a i. (Functor f, Prop a, Iso i) => i (Yoneda f a) (f a)
 lowerYoneda = inv' liftYoneda
 
 -- | Coyoneda encoding for a linear-logic 'Functor'.
 data Coyoneda f a where
-  Coyoneda :: Prop' r => (r ⊸ a) -> f r %1 -> Coyoneda f a
+  Coyoneda :: Prop r => (r ⊸ a) -> f r %1 -> Coyoneda f a
 
 -- | Refutations of 'Coyoneda'.
-newtype Cononeda f a = Cononeda (forall r. Prop' r => f r ⊸ WhyNot (a <#- r))
+newtype Cononeda f a = Cononeda (forall r. Prop r => f r ⊸ WhyNot (a <#- r))
 
-runCononeda :: forall r l a f. (Prop' r, Lol l) => Cononeda f a %1 -> l (f r) (WhyNot (a <#- r))
+runCononeda :: forall r l a f. (Prop r, Lol l) => Cononeda f a %1 -> l (f r) (WhyNot (a <#- r))
 runCononeda (Cononeda f) = fun f
 
-instance (Functor f, Prop' a) => Prop' (Coyoneda f a) where
+instance (Functor f, Prop a) => Prop (Coyoneda f a) where
   type Not (Coyoneda f a) = Cononeda f a
   Coyoneda r2a fr != k = because (runCononeda k fr) r2a
 
-instance (Functor f, Prop' a) => Prop' (Cononeda f a) where
+instance (Functor f, Prop a) => Prop (Cononeda f a) where
   type Not (Cononeda f a) = Coyoneda f a
   k != Coyoneda r2a fr = because (runCononeda k fr) r2a
 
-cononeda :: (forall r. Prop' r => f r ⊸ WhyNot (a <#- r)) %1 -> Cononeda f a
+cononeda :: (forall r. Prop r => f r ⊸ WhyNot (a <#- r)) %1 -> Cononeda f a
 cononeda = Cononeda
 
 -- | avoids the Functor constraint